Add assets for enemies
