--- src/config/phaserConfig.js
@@ -1,24 +1,0 @@
-// import Phaser from 'phaser';
-// import {BootScene} from './../scenes/BootScene'
-// // import WorldScene from './../scenes/WorldScene'
-
-// const config = {
-//   type: Phaser.AUTO,
-//   parent: 'app',
-//   width: 320,
-//   height: 240,
-//   zoom: 2,
-//   pixelArt: true,
-//   physics: {
-//     default: 'arcade',
-//     arcade: {
-//       gravity: { y: 0 }
-//     }
-//   },
-//   scene: [
-//     BootScene
-//     // WorldScene
-//   ]
-// }
-
-// export default config
