+++ src/scenes/WorldScene.js
@@ -1,23 +1,139 @@
-// import Phaser from 'phaser';
-// import map from './../assets/map/map.json'
+import Phaser from 'phaser';
 
-// export default class WorldScene extends Phaser.Scene {
-//   constructor() {
-//     super({ key: 'WorldScene' })
-//   }
 
-//   preload() {
+export default class WorldScene extends Phaser.Scene {
+  constructor() {
+    super({key: 'WorldScene'})
+  }
 
-//   }
+  preload(){
+        
+  }
 
-//   create() {
-//     var map = this.make.tilemap({ key: 'map' });
+  create() {
+    // create the map
+    var map = this.make.tilemap({ key: 'map' });
+        
+    // first parameter is the name of the tilemap in tiled
+    var tiles = map.addTilesetImage('spritesheet', 'tiles');
+        
+    // creating the layers
+    var grass = map.createStaticLayer('Grass', tiles, 0, 0);
+    var obstacles = map.createStaticLayer('Obstacles', tiles, 0, 0);
+        
+    // make all tiles in obstacles collidable
+    obstacles.setCollisionByExclusion([-1]);
+        
+    //  animation with key 'left', we don't need left and right as we will use one and flip the sprite
+    this.anims.create({
+        key: 'left',
+        frames: this.anims.generateFrameNumbers('player', { frames: [1, 2]}),
+        frameRate: 10,
+        repeat: -1
+    });
+        
+    // animation with key 'right'
+    this.anims.create({
+        key: 'right',
+        frames: this.anims.generateFrameNumbers('player', { frames: [1, 2 ] }),
+        frameRate: 10,
+        repeat: -1
+    });
 
-//     var tiles = map.addTilesetImage('spritesheet', 'tiles');
+    this.anims.create({
+        key: 'up',
+        frames: this.anims.generateFrameNumbers('player', { frames: [3, 4, 5]}),
+        frameRate: 10,
+        repeat: -1
+    });
 
-//     var grass = map.createStaticLayer('Grass', tiles, 0, 0);
-//     var obstacles = map.createStaticLayer('Obstacles', tiles, 0, 0);
-//     obstacles.setCollisionByExclusion([-1]);
-//     this.player = this.physics.add.sprite(50, 100, 'player', 6);
-//   }
-// }
+    this.anims.create({
+        key: 'down',
+        frames: this.anims.generateFrameNumbers('player', { frames: [ 7, 8 ] }),
+        frameRate: 10,
+        repeat: -1
+    });        
+
+    // our player sprite created through the phycis system
+    this.player = this.physics.add.sprite(50, 100, 'player', 6);
+    
+    // don't go out of the map
+    this.physics.world.bounds.width = map.widthInPixels;
+    this.physics.world.bounds.height = map.heightInPixels;
+    this.player.setCollideWorldBounds(true);
+    
+    // don't walk on trees
+    this.physics.add.collider(this.player, obstacles);
+
+    // limit camera to map
+    this.cameras.main.setBounds(0, 0, map.widthInPixels, map.heightInPixels);
+    this.cameras.main.startFollow(this.player);
+    this.cameras.main.roundPixels = true; // avoid tile bleed
+
+    // user input
+    this.cursors = this.input.keyboard.createCursorKeys();
+    
+    // where the enemies will be
+    this.spawns = this.physics.add.group({ classType: Phaser.GameObjects.Zone });
+    for(var i = 0; i < 30; i++) {
+        var x = Phaser.Math.RND.between(0, this.physics.world.bounds.width);
+        var y = Phaser.Math.RND.between(0, this.physics.world.bounds.height);
+        // parameters are x, y, width, height
+        this.spawns.create(x, y, 20, 20);
+    }        
+      // add collider
+      this.physics.add.overlap(this.player, this.spawns, this.onMeetEnemy, false, this);
+  }
+
+  onMeetEnemy(player, zone) {        
+      // we move the zone to some other location
+      zone.x = Phaser.Math.RND.between(0, this.physics.world.bounds.width);
+      zone.y = Phaser.Math.RND.between(0, this.physics.world.bounds.height);
+      
+      // shake the world
+      this.cameras.main.shake(300);
+      
+      // start battle 
+  }
+
+  update(time, delta) {
+    // this.controls.update(delta);
+    
+    this.player.body.setVelocity(0);
+
+    // Horizontal movement
+    if (this.cursors.left.isDown) {
+        this.player.body.setVelocityX(-80);
+    }
+    else if (this.cursors.right.isDown) {
+        this.player.body.setVelocityX(80);
+    }
+
+    // Vertical movement
+    if (this.cursors.up.isDown) {
+        this.player.body.setVelocityY(-80);
+    }
+    else if (this.cursors.down.isDown) {
+        this.player.body.setVelocityY(80);
+    }        
+
+    // Update the animation last and give left/right animations precedence over up/down animations
+    if (this.cursors.left.isDown) {
+        this.player.anims.play('left', true);
+        this.player.flipX = false;
+    }
+    else if (this.cursors.right.isDown) {
+        this.player.anims.play('right', true);
+        this.player.flipX = true;
+    }
+    else if (this.cursors.up.isDown) {
+        this.player.anims.play('up', true);
+    }
+    else if (this.cursors.down.isDown) {
+        this.player.anims.play('down', true);
+    }
+    else {
+        this.player.anims.stop();
+    }
+  }
+}
